feat(advanced-credit-calculator): generalise advanced credit calculator to IB courses
