+++ src/data/nonOfferedCourses.json
@@ -7,16 +7,6 @@
     { "id": "03-110", "name": "AP 5 Biology", "units": "9" },
 
     { "id": "21-111", "name": "Calculus (AP Calculus AB)", "units": "10" },
-    {
-        "id": "21-120",
-        "name": "Differential and Integral Calculus",
-        "units": "10"
-    },
-    {
-        "id": "21-122",
-        "name": "Integrations, Differential Equations and Approximation",
-        "units": "10"
-    },
 
     {
         "id": "99-005",
@@ -33,9 +23,6 @@
 
     { "id": "82-011", "name": "AP 4 Chinese", "units": "12" },
     { "id": "82-012", "name": "AP 5 Chinese", "units": "12" },
-
-    { "id": "15-110", "name": "Principles of Computing", "units": "10" },
-    { "id": "15-112", "name": "Fundamentals of Programming", "units": "12" },
 
     { "id": "60-012", "name": "AP Drawing", "units": "9" },
     { "id": "73-011", "name": "AP Economics (Micro)", "units": "9" },
@@ -77,17 +64,6 @@
 
     { "id": "57-012", "name": "AP Music Theory", "units": "9" },
 
-    {
-        "id": "33-142",
-        "name": "Physics II for Engineering Students",
-        "units": "12"
-    },
-    {
-        "id": "33-141",
-        "name": "Physics I for Engineering Students",
-        "units": "12"
-    },
-
     { "id": "85-011", "name": "AP Psychology", "units": "9" },
     { "id": "79-016", "name": "AP Anthropology", "units": "9" },
 
@@ -111,8 +87,54 @@
         "units": "9"
     },
 
-    { "id": "36-200", "name": "Reasoning with Data", "units": "9" },
-
     { "id": "79-012", "name": "AP United States History", "units": "9" },
-    { "id": "79-015", "name": "AP World History: Modern", "units": "9" }
+    { "id": "79-015", "name": "AP World History: Modern", "units": "9" },
+
+    { "id": "82-025", "name": "IB 6 Chinese", "units": "12"},
+    { "id": "82-026", "name": "IB 7 Chinese", "units": "12"},
+
+    { "id": "66-021", "name": "IB Classical Languages: Greek", "units": "9" },
+    { "id": "66-022", "name": "IB 7 Classical Languages: Greek", "units": "9" },
+
+    { "id": "66-023", "name": "IB Classical Languages: Latin", "units": "9" },
+    { "id": "66-024", "name": "IB 7 Classical Languages: Latin", "units": "9" },
+
+    { "id": "51-013", "name": "IB Design Technology", "units": "9" },
+
+    { "id": "76-013", "name": "IB English A Language & Lit", "units": "9" },
+    { "id": "76-014", "name": "IB English A Literature", "units": "9" },
+
+    { "id": "76-017", "name": "IB Film", "units": "9" },
+
+    { "id": "82-027", "name": "IB French A", "units": "9" },
+    { "id": "82-028", "name": "IB 7 French A", "units": "9" },
+
+    { "id": "66-012", "name": "IB Geography", "units": "9" },
+
+    { "id": "82-029", "name": "IB 6 German", "units": "9" },
+    { "id": "82-030", "name": "IB 7 German", "units": "9" },
+
+    { "id": "84-013", "name": "IB Global Politics", "units": "9" },
+    { "id": "79-014", "name": "IB History", "units": "9" },
+    { "id": "67-011", "name": "IB Information Technology in a Global Society (ITGS)", "units": "9" },
+    { "id": "79-013", "name": "IB Islamic History", "units": "9" },
+
+    { "id": "82-031", "name": "IB 6 Italian", "units": "9" },
+    { "id": "82-032", "name": "IB 7 Italian", "units": "9" },
+
+    { "id": "82-033", "name": "IB 6 Japanese", "units": "12" },
+    { "id": "82-034", "name": "IB 7 Japanese", "units": "12" },
+
+
+    { "id": "57-013", "name": "IB Music", "units": "9" },
+    { "id": "80-011", "name": "IB Philosophy", "units": "9" },
+    { "id": "85-012", "name": "IB Psychology", "units": "9" },
+
+    { "id": "82-035", "name": "IB 6 Russian", "units": "12" },
+    { "id": "82-036", "name": "IB 7 Russian", "units": "12" },
+
+    { "id": "79-016", "name": "IB Anthropology", "units": "9" },
+    
+    { "id": "82-037", "name": "IB 6 Spanish", "units": "9" },
+    { "id": "82-038", "name": "IB 7 Spanish", "units": "9" }
 ]
